Added explicit cast from size_t to int32.

+++ rclcpp/executors/cbg_executor/src/examples_rclcpp_cbg_executor/ping_node.cpp
@@ -45,7 +45,7 @@ PingNode::PingNode()
 void PingNode::send_ping()
 {
   std_msgs::msg::Int32 msg;
-  msg.data = rtt_data_.size();
+  msg.data = static_cast<int32_t>(rtt_data_.size());
   rtt_data_.push_back(RTTData(now()));
   high_ping_publisher_->publish(msg);
   low_ping_publisher_->publish(msg);
